Add timeout to flush buffer if last flush was longer than x seconds a… (#49)
Add timeout to flush buffer if last flush was longer than x seconds ago.
When KeyboardInterrupt is raised at the end of test_buffer_timeout() it will flush the buffer, this breaks the test logic causing it to always return true. Added except SystemExit to Worker.consume() to provide a way to break out of the consumer loop without flushing the buffer, SystemExit is now raised after the test.